+++ packages/tm-core/src/repositories/supabase/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Supabase repository implementations
+ */
+export { SupabaseTaskRepository } from './supabase-task-repository.js';
+export { DependencyFetcher, type DependencyWithDisplayId } from './dependency-fetcher.js';
